requirements.txt
fastapi>=0.103.2
uvicorn[standard]>=0.40.0
sqlmodel>=0.0.30
python-jose[cryptography]>=3.5.0
passlib>=1.7.4
pyotp>=2.9.0
qrcode[pil]>=7.4.0
pydantic[email]>=1.10.0
email-validator>=2.0.0
python-dotenv>=1.0.0
